added inline and shellconf, more impl of isys
Added inlines header. The library consists only of very short, inlined
functions and macros.

Added ShellConf. It is the POSIX-compatible shell script for building simple C
programs such as this library.

Improved the coverage of ISys. Partway to the goal of having a basic threading
and synchronization library for Windows and POSIX.