added lexer, fixed shellconf, bug fixes
Add a lexer library. It is a simple way for parsing lexemes from an input. It
still needs work before being completely usable.

Fixed bugs in ShellConf when dealing with projects without autogenerated
`inc.h` files.

Fixed a bug in Memory Debugger that caused a segfault in the `vmprintf`
function.