added betterio, strprefix function, bug fix in avl
Added the BetterIO library. The library provides an alternative set of
functions for read and writing to files. Probably the most useful aspect is
the ability to use callbacks for writing to a file -- essentially calling a
function to write to an output as needed. This makes writing custom printing
functions much easier.

Fixed a bug in AVL Trees that used a non-const pointer. This bug basically
forced client code to use non-const pointers.

Added `strprefix` to Inlines Header. As the name says, the `strprefix`
function determines if a a string is a prefix.