improved debug info on lexer, mdbg, and emech
Fixed the debugging info on the Lexer. I fixed the lexer code to assign the
appropriate line and column information for tokens. Also, I added error
formatting functions for tokens and the head token for the lexer.

Added a version of `fatal` with line information to the Memory Debugger. If
`RELEASE` is not defined, then the call to `fatal` redirects to `fatal_debug`
with the caller's file and line information.

Added a fail with message function `failmsg` to EMech. The added function
`failmsg` calls the error handler and returns the heap allocated parameter.

Fixed `clean` rule for ShellConf. I added the dependency files (`*.d`) to the
`CLEAN` variable.